\name{annotation_file}
\alias{annotation_file}
\docType{data}
\title{annotation file for the 450k probes}
\description{UCSC annotation for the 450k DNA methylation probes. The annotation was obtained from "Illumina GPL13534_HumanMethylation450_15017482_v1.1" file with few amendments on the gene names}
\usage{data("annotation_file")}
\format{A data frame}
\keyword{datasets}
