\name{test_data}
\alias{test_data}
\docType{data}
\title{test dataset of 450k DNA methylation}
\description{This is a small dataset of 450k DNA methylation array with 10000 probes. The dataset has four columns; columns 1 and 2 contain normalised beta values for paediatric B ALL cases, and columns 3 and 4 contain beta values for controls (remission cases)}
\usage{data("test_data")}
\format{A data frame}
\details{a small test dataset}
\keyword{datasets}
\references{Busche S, Ge B, Vidal R, etc. Integration of high-resolution methylome and transcriptome analyses to dissect epigenomic changes in childhood acute lymphoblastic leukaemia. Cancer Research 2013; 73(14); 4323-4336}